rename files and arrange folder structure, remove initial loader, add favicon
