+++ src/main/webapp/app/modules/dashboard/main-dashboard.page.tsx
@@ -3,7 +3,7 @@ import React, { useEffect } from 'react';
 import { Storage, translate } from "react-jhipster";
 import { useNavigate } from 'react-router-dom';
 import { useAppDispatch } from "app/config/store";
-import { reset } from 'app/modules/new-profile/create-new-profile/pages/create-new-file/createNewProfile.reducer';
+import { reset } from 'app/modules/new-profile/create-new-file/createNewProfile.reducer';
 
 const MainDashboardPage = () => {
