+++ src/main/webapp/app/modules/new-profile/legal-bonds-page/components/legalBonds.reducer.ts
