+++ src/main/webapp/app/modules/new-profile/legal-bonds-page/components/shared/accountStatement.component.tsx
@@ -11,7 +11,7 @@ import {
   useForm,
 } from "react-hook-form";
 import { CurrencyList } from "app/modules/shared/constants";
-import LoaderComponent from "app/modules/shared/loaderComponent/loaderComponent";
+import LoaderComponent from "app/shared/components/loaderComponent/loaderComponent";
 import { addLegalBond } from "../legalBonds.reducer";
 
 const AccountStatement = (props) => {
