+++ src/main/webapp/app/modules/new-profile/legal-bonds-page/components/shared/defendantInfo/shared/draftTable.component.tsx
