+++ src/main/webapp/app/modules/new-profile/legal-bonds-page/components/shared/defendantInfo/shared/invoiceTable.component.tsx
