+++ src/main/webapp/app/modules/new-profile/legal-bonds-page/components/shared/defendantInfo/shared/statementTable.component.tsx
