+++ src/main/webapp/app/modules/new-profile/legal-bonds-page/components/shared/mortgageBond.component.tsx
@@ -10,7 +10,7 @@ import { translate, Storage } from "react-jhipster";
 import { useAppDispatch, useAppSelector } from "app/config/store";
 import { useForm } from "react-hook-form";
 import { CurrencyList } from "app/modules/shared/constants";
-import LoaderComponent from "app/modules/shared/loaderComponent/loaderComponent";
+import LoaderComponent from "app/shared/components/loaderComponent/loaderComponent";
 import dayjs from "dayjs";
 import { addLegalBond } from "../legalBonds.reducer";
