+++ src/main/webapp/app/modules/new-profile/legal-bonds-page/components/shared/writtenAcknowledgmentTrustBond.component.tsx
