+++ src/main/webapp/app/modules/new-profile/select-file-type/select-file-type.reducer.ts
