+++ src/main/webapp/app/modules/new-profile/select-file-type/shared/collection.component.tsx
