+++ src/main/webapp/app/modules/new-profile/select-file-type/shared/companySubFileData.component.tsx
