+++ src/main/webapp/app/shared/components/breadcrumbs.Component/breadcrumb.Component.scss
@@ -1,4 +1,4 @@
-@import "../../../content/scss/variables/bgAndColor.scss";
+@import "../../../../content/scss/variables/bgAndColor.scss";
 
 .breadcrumbComponent {
   margin: 0 0 32px;
