+++ src/main/webapp/app/shared/components/loaderComponent/loaderComponent.scss
