+++ src/main/webapp/index.html
@@ -8,9 +8,9 @@
     <meta name="google" content="notranslate" />
     <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no" />
     <meta name="theme-color" content="#000000" />
-    <link rel="icon" href="favicon.ico" />
+    <link rel="icon" href="content/images/logo.svg" />
     <link rel="manifest" href="manifest.webapp" />
-    <link rel="stylesheet" href="content/css/loading.css" />
+    <!-- <link rel="stylesheet" href="content/css/loading.css" /> -->
     <!-- jhipster-needle-add-resources-to-root - JHipster will add new resources here -->
   </head>
   <body>
@@ -21,7 +21,7 @@
       </p>
     <![endif]-->
     <div id="root">
-      <div class="app-loading">
+      <!-- <div class="app-loading">
         <div class="lds-pacman">
           <div>
             <div></div>
@@ -34,7 +34,7 @@
             <div></div>
           </div>
         </div>
-      </div>
+      </div> -->
       <div class="app-loading">
         <div id="jhipster-error" style="display: none">
           <!-- This content is for troubleshooting purpose and will be removed when app renders -->
